Change function header to return vector of doubles for Findlightsource vector.

+++ orientation_reflectance/programs.h
@@ -29,7 +29,7 @@ namespace Programs {
   //Finds the x,y,z compnent of the light source vector for the given image
   //@param an_image, an image with one light source
   //@return, vector of x,y,z component of light source.
-	vector<int> FindLightSourceVector(Image* an_image, int centerX, int centerY, int radius);
+	vector<double> FindLightSourceVector(Image* an_image, int centerX, int centerY, int radius);
 
   //Finds the surface normals and applies it to image1
   //@param directions, a vector containing light source directions for each of the three images. contains X,Y,Z component
